+++ packages/featureserver/src/index.js
@@ -2,6 +2,7 @@ const info = require('./info.js')
 
 module.exports = {
   route: require('./route.js'),
+  restInfo: info.restInfo,
   serverInfo: info.serverInfo,
   layerInfo: info.layerInfo,
   layersInfo: info.layersInfo,