+++ packages/featureserver/src/templates.js
@@ -5,13 +5,14 @@ const { computeFieldObject, createFieldAliases, createStatFields } = require('./
 const { computeSpatialReference, computeExtent } = require('./geometry')
 const { createClassBreakInfos, createUniqueValueInfos } = require('./generateRenderer/createClassificationInfos')
 
-module.exports = { renderLayer, renderFeatures, renderStatistics, renderServer, renderStats, renderClassBreaks, renderUniqueValue }
+module.exports = { renderRestInfo, renderLayer, renderFeatures, renderStatistics, renderServer, renderStats, renderClassBreaks, renderUniqueValue }
 
 const templates = {
-  layer: require('../templates/layer.json'),
+  layer: Object.assign(require('../templates/layer.json'), require('../templates/version.json')),
   features: require('../templates/features.json'),
   statistics: require('../templates/statistics.json'),
-  server: require('../templates/server.json'),
+  restInfo: Object.assign(require('../templates/rest-info.json'), require('../templates/version.json')),
+  server: Object.assign(require('../templates/server.json'), require('../templates/version.json')),
   field: require('../templates/field.json'),
   objectIDField: require('../templates/oid-field.json')
 }
@@ -85,6 +86,15 @@ function renderStatistics (featureCollection = {}, options = {}) {
 
   if (json.fields) json.fields = computeFieldObject(data, 'statistics', options)
   if (json.features) json.features = data.features
+  return json
+}
+
+/**
+ * Get and return the templated rest/info response
+ * @param {*} restInfo
+ */
+function renderRestInfo () {
+  const json = _.cloneDeep(templates.restInfo)
   return json
 }